+++ hephaistos-data/lua/Filters/CombatPresentation.lua
@@ -1,18 +1,3 @@
--- boiling blood animation when getting hit by [Redacted] bloodstone
-Hephaistos.CreateAnimation[CastEmbeddedPresentationStart] = function(params)
-	return Hephaistos.MatchAll(params, { Name = "HadesBloodstoneVignette", DestinationId = ScreenAnchors.HadesBloodstoneVignette })
-end
-
--- poison animation when getting hit by poison
-Hephaistos.CreateAnimation[StartStyxPoisonPresentation] = function(params)
-	return Hephaistos.MatchAll(params, { Name = "PoisonVignetteLoop", DestinationId = ScreenAnchors.PoisonVignette })
-end
-
--- fire animation when getting hit by lava
-Hephaistos.SetAnimation[StartLavaPresentation] = function(params)
-	return Hephaistos.MatchAll(params, { Name = "LavaVignetteSpawner", DestinationId = ScreenAnchors.LavaVignette })
-end
-
 -- fix full screen alert overlay displace fx
 -- for all calls e.g. call, [Redacted] call, Theseus call and Cerberus interaction during [Redacted] fight
 local function replaceObstacleAlertFx(params)