+++ hephaistos-data/lua/Filters/EventPresentation.lua
@@ -1,12 +1,3 @@
--- vignette flashes when getting hit
-Hephaistos.CreateAnimation[HeroDamagePresentationThread] = function(params)
-	return Hephaistos.MatchAll(params, { Name = "BloodFrame", UseScreenLocation = true, OffsetX = ScreenCenterX, OffsetY = ScreenCenterY, GroupName = "Vignette" })
-end
-
-Hephaistos.CreateAnimation[HeroArmorDamagePresentationThread] = function(params)
-	return Hephaistos.MatchAll(params, { Name = "BloodFrame", UseScreenLocation = true, OffsetX = ScreenCenterX, OffsetY = ScreenCenterY })
-end
-
 -- fix full screen alert overlay displace fx
 -- for important events such as companion quests (Achilles, Sisyphus, etc.) or hidden aspects reveal
 local function filterAnimationAlertFx(params)