+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+-   Fix text overflowing out of GUI components.
+-   Properly scale vignettes displayed when hit (blood frame, boiling blood, poison, lava).
 -   Re-fix world map shown between regions (oops :D).
 -   Fix full screen displacement FX for various occasions (calls, Chaos interact, Hades speaking, etc.).
 -   Fix assist/summon (e.g. Sisyphus) overlay.